feat(tool): add thread-safe whitelist management and dynamic desc (#379)
## AgentScope-Java Version

1.0.4

## Description

Add a list of allowed commands to the description sent to LLM. This
helps to reduce the occurrence of LLM running commands that are not
allowed. Although it incurs an increase in the consumption of one-time
tokens, it can reduce the number of tool invocations, which is a more
token-consuming behavior. At the same time, make the modification of the
allowedCommand set thread-safe.

## Checklist

Please check the following items before code is ready to be reviewed.

- [x]  Code has been formatted with `mvn spotless:apply`
- [x]  All tests are passing (`mvn test`)
- [x]  Javadoc comments are complete and follow project conventions
- [x] Related documentation has been updated (e.g. links, examples,
etc.)
- [x]  Code is ready for review

---------

Co-authored-by: Albumen Kevin <jhq0812@gmail.com>